changed setup command
