+++ src/classes/utils.tsx
@@ -983,7 +983,7 @@ export default class Utils {
 
       let green = ` ${title}`.padEnd(25," ")
       let white = ` Perri's brewery edition `.padEnd(25," ")
-      let red = ` v${pjson.version}.${type} `.padStart(25," ")
+      let red = ` v${pjson.version} ${type} `.padStart(25," ")
 
       return chalk.bgGreen.whiteBright(green) +
          chalk.bgWhite.blue(white) +
