+++ src/krill/components/title.tsx
@@ -30,7 +30,7 @@ export default function Title({ title="", version=""}) : JSX.Element {
 
    let green = ` ${title}`.padEnd(25," ")
    let white = ` Perri's brewery edition `.padEnd(25," ")
-   let red = ` v${pjson.version}.${type}  `.padStart(25," ")
+   let red = ` v${pjson.version} ${type}  `.padStart(25," ")
 
    return(
       <>
